// sanity/sanity-types.ts
export type Author = {
  _id: string;
  _type: "author";
  _createdAt: string;
  _updatedAt: string;
  _rev: string;
  id?: string;
  name?: string;
  familyName?: string;
  email?: string;
  image?: string;
};